requirements.txt
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.7.0
pillow>=9.0.0
matplotlib>=3.5.0
pandas>=1.4.0
albumentations>=1.3.0
PyYAML>=6.0
tqdm>=4.65.0
scikit-learn>=1.0.0
wandb>=0.15.0  # for experiment tracking
